+++ myems_historical_db.sql
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS `myems_historical_db`.`tbl_energy_value` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `point_id` BIGINT NOT NULL,
   `utc_date_time` DATETIME NOT NULL,
-  `actual_value` DOUBLE NOT NULL,
+  `actual_value` DECIMAL(18, 3) NOT NULL,
   `is_bad` BOOL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_energy_value_index_1` ON  `myems_historical_db`.`tbl_energy_value`  (`point_id` , `utc_date_time`);
